requirements.txt
fastapi
uvicorn
langchain-google-genai
pydantic
python-dotenv
GitPython